+++ CHANGELOG.md
@@ -8,6 +8,10 @@
 
 - Improved `to_pandas` to persist the original timezone offset for TIMESTAMP_TZ type.
 
+#### New Features
+
+- Added support for `TimedeltaIndex.mean` method.
+
 ## 1.22.0 (2024-09-10)
 
 ### Snowpark Python API Updates
@@ -90,7 +94,6 @@
   - support for `diff` with timestamp columns on `axis=0` and `axis=1`
   - support for `TimedeltaIndex` methods: `ceil`, `floor` and `round`.
   - support for `TimedeltaIndex.total_seconds` method.
-  - support for `TimedeltaIndex.mean` method.
 - Added support for index's arithmetic and comparison operators.
 - Added support for `Series.dt.round`.
 - Added documentation pages for `DatetimeIndex`.