+++ CHANGELOG.md
@@ -8,6 +8,9 @@
 
 - Improved `to_pandas` to persist the original timezone offset for TIMESTAMP_TZ type.
 
+#### Bug Fixes
+- Fixed a bug where an `Index` object created from a `Series`/`DataFrame` incorrectly updates the `Series`/`DataFrame`'s index name when it is not supposed to.
+
 ## 1.22.0 (2024-09-10)
 
 ### Snowpark Python API Updates